refactor: change SelectField
